updated and faster starship config

+++ misc/update.sh
@@ -24,7 +24,7 @@ git submodule update --init --recursive --remote
 
 echo "\n\nrefreshing dotfiles links .."
 python3 ../bin/ghar install
-ln -s ~/ghar/dotfiles/misc/starship.toml ~/.config
+[[ -e ~/.config/starship.toml ]] || ln -s ~/ghar/dotfiles/misc/starship.toml ~/.config
 
 command -v starship >/dev/null 2>&1 || curl -fsSL https://starship.rs/install.sh | bash
